third_party/blink/web_tests/external/wpt/accname/name/comp_name_from_heading.tentative-expected.txt

This is a testharness.js-based test.
Harness Error. harness_status.status = 1 , harness_status.message = Uncaught Element should have attribute 'data-expectedlabel'. Element: <div role="group" data-expectedlabel="" data-testname="group role, verify name is NOT from heading" class="ex-label">\n  <div role="group">\n    <h3>Not this one</h3>\n  </div>\n  <h2>nor this one</h2>\n  <p>Nor this content.</p>\n</div>
[FAIL] alertdialog role, name from heading
  assert_equals: <div role="alertdialog" data-expectedlabel="alertdialog role heading contents" data-testname="alertdialog role, name from heading" class="ex-label">\n  <h2>alertdialog role heading contents</h2>\n  <p>More alertdialog content.</p>\n</div> expected "alertdialog role heading contents" but got ""
[FAIL] article role, name from heading
  assert_equals: <div role="article" data-expectedlabel="article role heading contents" data-testname="article role, name from heading" class="ex-label">\n  <h2>article role heading contents</h2>\n  <p>More article content.</p>\n</div> expected "article role heading contents" but got ""
[FAIL] dialog role, name from heading
  assert_equals: <div role="dialog" data-expectedlabel="dialog role heading contents" data-testname="dialog role, name from heading" class="ex-label">\n  <h2>dialog role heading contents</h2>\n  <p>More dialog content.</p>\n</div> expected "dialog role heading contents" but got ""
Harness: the test ran to completion.
